+++ examples/complete/versions.tf
@@ -7,6 +7,16 @@ terraform {
       version = ">= 5.0.0"
     }
 
+    time = {
+      source  = "hashicorp/time"
+      version = ">= 0.9"
+    }
+
+    archive = {
+      source  = "hashicorp/archive"
+      version = ">= 2.2.0"
+    }
+
     random = {
       source  = "hashicorp/random"
       version = ">= 3.1.0"