+++ CHANGELOG.md
@@ -5,9 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- fix: `Error Putting S3 Notification Configuration: Unable to validate the following destination configurations` when using using notifications on complete example.
 - fix: replication configuration showing drift detection when filter is used, even when values are unchanged
 - fix: CKV_AWS_300 #Ensure S3 lifecycle configuration sets period for aborting failed uploads
+- fix: CKV2_AWS_65: #Ensure access control lists for S3 buckets are disabled
 - feat: show notification usage in complete example after fixing the notification error
 - feat: Document on README needed permissions and steps to replicate encrypted bucket objects
 - feat: Acceleration Configuration
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: S3 Object Copy
 - feat: S3 Object
 - fix: CKV_TF_1 #"Ensure Terraform module sources use a commit hash"
+
+## [2.3.1] - 2023-09-26
+- fix: `Error Putting S3 Notification Configuration: Unable to validate the following destination configurations` when using using notifications on complete example.
+- feat: show full module coverage in complete examples for bucket notification feature.
 
 ## [2.3.0] - 2023-09-07
 - feat: Added Lifecycle Configuration which fixes checkov alert `CKV2_AWS_62`
@@ -75,8 +79,9 @@ The following features have now been added as stand-alone resources in this rele
 - Initial commit
 - Included most basic/common settings in module
 
-[Unreleased]: https://github.com/boldlink/terraform-aws-s3/compare/2.3.0...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-aws-s3/compare/2.3.1...HEAD
 
+[2.3.1]: https://github.com/boldlink/terraform-aws-s3/releases/tag/2.3.1
 [2.3.0]: https://github.com/boldlink/terraform-aws-s3/releases/tag/2.3.0
 [2.2.1]: https://github.com/boldlink/terraform-aws-s3/releases/tag/2.2.1
 [2.2.0]: https://github.com/boldlink/terraform-aws-s3/releases/tag/2.2.0